flask/requirements.txt
dataclasses
gpu-info
joblib
numpy
onnxruntime-gpu
Pillow
protobuf
psutil
pygobject
scikit-learn
scipy
six
threadpoolctl
torch
torchvision
typing-extensions